Character List
Added a Character List which will display all Characters.

Before this the process of searching Characters was to either go through them individually using the Account Info Form or the Players Online Tab.

Also has a Filter By Item UID/Name box which allows you to quickly identify the owner of an Item.